+++ src/main/java/com/feiyu/discord/sg/tavern/listeners/MemberJoinListener.java
@@ -3,12 +3,12 @@ package com.feiyu.discord.sg.tavern.listeners;
 import com.feiyu.discord.sg.tavern.config.ValuesConfig;
 import com.feiyu.discord.sg.tavern.entities.NewJoinerEntity;
 import com.feiyu.discord.sg.tavern.repositories.NewJoinerRepository;
+import com.feiyu.discord.sg.tavern.services.MessageService;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import net.dv8tion.jda.api.entities.Guild;
 import net.dv8tion.jda.api.entities.Role;
 import net.dv8tion.jda.api.entities.User;
-import net.dv8tion.jda.api.entities.channel.concrete.TextChannel;
 import net.dv8tion.jda.api.events.guild.member.GuildMemberJoinEvent;
 import net.dv8tion.jda.api.hooks.ListenerAdapter;
 import org.springframework.stereotype.Component;
@@ -22,6 +22,7 @@ public class MemberJoinListener extends ListenerAdapter {
     
     private final ValuesConfig valuesConfig;
     private final NewJoinerRepository newJoinerRepository;
+    private final MessageService messageService;
     
     @Override
     public void onGuildMemberJoin(GuildMemberJoinEvent event) {
@@ -37,12 +38,12 @@ public class MemberJoinListener extends ListenerAdapter {
         
         newJoinerRepository.save(newJoinerEntity);
         
-        log.info("Someone joined and assigned newcomer role : {} - {} ", user.getEffectiveName(), user.getName());
-        guild.addRoleToMember(user, newJoinerRole).queue();
-        
         String adminMessage = "Someone has joined the server : " + user.getEffectiveName() + " - " + user.getName();
-        TextChannel adminChannel = guild.getTextChannelById(valuesConfig.getAdminBotChannelId());
-        adminChannel.sendMessage(adminMessage).queue();
+        messageService.sendAdminChannelMessage(guild, adminMessage);
+        
+        String adminRoleMessage = user.getEffectiveName() + " - " +  user.getName() + " assigned role : " + newJoinerRole.getName();
+        guild.addRoleToMember(user, newJoinerRole).queue();
+        messageService.sendAdminChannelMessage(guild, adminRoleMessage);
     }
     
 }